Update documentation and refactoring
