+++ internal/char/char.go
