+++ internal/char/char_test.go
