+++ internal/command/command.go
@@ -183,8 +183,6 @@ func (c *Command) Run(args []string) error {
 		return err
 	}
 
-	// defer f.Close()
-
 	spec, err := c.funcs.Parse(filename, f)
 	if err != nil {
 		return err
