+++ internal/regex/parser/ast/ast.go
@@ -13,7 +13,7 @@ import (
 	"github.com/moorara/algo/list"
 	"github.com/moorara/algo/parser/combinator"
 
-	"github.com/gardenbed/emerge/internal/regex/char"
+	"github.com/gardenbed/emerge/internal/char"
 	"github.com/gardenbed/emerge/internal/regex/parser"
 )
