+++ internal/regex/parser/nfa/nfa.go
@@ -11,7 +11,7 @@ import (
 	"github.com/moorara/algo/automata"
 	"github.com/moorara/algo/parser/combinator"
 
-	"github.com/gardenbed/emerge/internal/regex/char"
+	"github.com/gardenbed/emerge/internal/char"
 	"github.com/gardenbed/emerge/internal/regex/parser"
 )
